Changed the way things are passed to Simplify at the end so that we can only have
	Simplify(X > Y)
	Simplify(X >= Y) or
	Simplify(X == Y)

other types of expressions get converted into negations of these. This little change actually
seems to get us a lot of mileage. For instance, now we have

In> LogicSimplify(A>3 Or B Or A<=3)
Out> A-3>0 Or B Or Not(A-3>0);

and so now

In> CanProve(A>3 Or B Or A<=3)
Out> True;

Furtermore, now we only need to worry about these 3 types of expressions in CanProve, not 5.