changes to follow recent rewrites of core function API.
- the refprog chapter yacasapi.chapt is now generated from corefunctions.h
and not from yacasapi.cpp. This is now generated by the C++ program
gencorefunctions and not by a perl script. The perl script api2txt.pl that
was tuned to yacasapi.cpp is removed.

- the refprog chapter on listing of core functions now gives somewhat more
detail about the functions (# of arguments, macro/nonmacro).

- the references to yacasapi.cpp in the docs are now replaced by references
to corefunctions.h.

- OPERATOR definitions moved to corefunctions.h from yacasapi.cpp; OPERATOR
macro defined in the three places where corefunctions.h is #include'd.