Made changes to allow compilation with gcc 3.4 (patch sent in by Jitse). Most changes had already been made a while back, for gcc 3.3 I believe. Some changes are odd, I don't understand why it is necessary.
One is not really a showstopper. It dealt with the OPERATOR and CORE_KERNEL_FUNCTION macros. Essentially, the choice was between

	#define M(x) g(x);
	...
	...
	M(x)


and

	#define M(x) g(x)
	...
	...
	M(x);

The choice between these two had not been consistent in the code, and older versions of gcc didn't complain about it. The choice has been made now for the latter.

Ayal