Removed LazyGlobal in favor of SetGlobalLazyVariable. This thus wraps two operations in one. This happened always any way, the variable was assigned, and set to be a lazy global.
This was initiated by the need to be able to flag warnings in the debug build
when a global variable was set. They are obviously intended to be global if
SetGlobalLazyVariable is called explicitly. This is more to protect against
the Set function accidentally setting a global variable in stead of a local variable.